+++ src/app/components/lineage-graph/ligeage-graph.settings.ts
@@ -1,6 +1,7 @@
 import { MiniMapPosition } from "@swimlane/ngx-graph";
 
-export const lineageConfig = {
+export const lineageConfig: LineageGraphConfig = {
+    layout: "dagre",
     draggingEnabled: false,
     panningEnabled: true,
     zoomEnabled: true,
@@ -11,10 +12,12 @@ export const lineageConfig = {
     autoZoom: true,
     autoCenter: true,
     showMiniMap: true,
+    nodeHeight: 70,
     miniMapPosition: MiniMapPosition.UpperLeft,
 };
 
 export interface LineageGraphConfig {
+    layout: string;
     draggingEnabled: boolean;
     panningEnabled: boolean;
     zoomEnabled: boolean;
@@ -25,5 +28,6 @@ export interface LineageGraphConfig {
     autoZoom: boolean;
     autoCenter: boolean;
     showMiniMap: boolean;
+    nodeHeight: number;
     miniMapPosition: MiniMapPosition;
 }