+++ src/app/dataset-view/additional-components/lineage-component/services/lineage-graph-builder.service.spec.ts
@@ -1,9 +1,8 @@
 import { TestBed } from "@angular/core/testing";
-
 import { LineageGraphBuilderService } from "./lineage-graph-builder.service";
 import { DatasetSubscriptionsService } from "src/app/dataset-view/dataset.subscriptions.service";
-import { mockLineageGraphUpdate } from "../../data-tabs.mock";
-import { LineageGraph } from "../lineage-model";
+import { mockBuildGraphNodesResult, mockLineageGraphUpdate } from "../../data-tabs.mock";
+import { LineageGraph, LineageGraphNodeData, LineageGraphNodeKind } from "../lineage-model";
 import { DatasetLineageBasicsFragment } from "src/app/api/kamu.graphql.interface";
 
 describe("LineageGraphBuilderService", () => {
@@ -32,6 +31,24 @@ describe("LineageGraphBuilderService", () => {
             expect(data).toBeDefined();
             expect(data.links.length).toEqual(4);
             expect(data.nodes.length).toEqual(5);
+        });
+    });
+
+    it("should check source nodes length", () => {
+        service.buildGraph().subscribe((data: LineageGraph) => {
+            const dataSource = mockLineageGraphUpdate.nodes.filter(
+                (node) => node.metadata.currentSource?.fetch.__typename === "FetchStepUrl",
+            );
+            const graphSourceNodes = data.nodes.filter(
+                (node) => (node.data as LineageGraphNodeData).kind === LineageGraphNodeKind.Source,
+            );
+            expect(dataSource.length).toEqual(graphSourceNodes.length);
+        });
+    });
+
+    it("should check nodes was built correctly", () => {
+        service.buildGraph().subscribe((data: LineageGraph) => {
+            expect(data.nodes).toEqual(mockBuildGraphNodesResult);
         });
     });
 });